Remove some warnings and improve debug.

+++ src/git.cpp
@@ -1581,7 +1581,7 @@ const QString Git::getLocalDate(SCRef gitDate) {
         return localDate;
 
     QDateTime d;
-    d.setTime_t(gitDate.toULong());
+    d.setTime_t(gitDate.toUInt());
     localDate = d.toString(Qt::LocalDate);
     localDates[gitDate] = localDate;
     return localDate;
@@ -2180,10 +2180,10 @@ void Git::on_loaded(FileHistory* fh, ulong byteSize, int loadTime,
 
             fh->loadTime += loadTime;
 
-            uint kb = byteSize / 1024;
-            float mbs = (float)byteSize / fh->loadTime / 1000;
+            ulong kb = byteSize / 1024;
+            double mbs = (double)byteSize / fh->loadTime / 1000;
             QString tmp;
-            tmp.sprintf("Loaded %i revisions  (%i KB),   "
+            tmp.sprintf("Loaded %i revisions  (%li KB),   "
                         "time elapsed: %i ms  (%.2f MB/s)",
                         fh->revs.count(), kb, fh->loadTime, mbs);
 
@@ -2629,7 +2629,7 @@ void Git::flushFileNames(FileNamesLoader& fl) {
     QVector<int>& dirs = fl.rfDirs;
 
     b.clear();
-    b.resize(2 * dirs.size() * sizeof(int));
+    b.resize(2 * dirs.size() * static_cast<int>(sizeof(int))); // static_cast avoids warning, (sizeof returns ulong on 64bits)
 
     int* d = (int*)(b.data());
